requirements.txt
django
djangorestframework

asgiref
pip
setuptools
sqlparse
tzdata
